Allow indirection
This pull-request updates #16, by allowing instructions to be parsed
which operate upon indirection.

For example:

        mov rax, [rbx]

Or:

        inc byte ptr [rax]
        inc word ptr [rax]
        inc qword ptr [rax]

Those instructions are not yet encoded, but they will be parsed.

We recognize two flags:

* Indirection
  * When an address is involved.
* Size
  * For the size which we're operating upon
    * `byte ptr` -> 8-bits
    * `word ptr` -> 16-bits
    * `dword ptr` -> 32-bits
    * `qword ptr` -> 64-bits

(Of course we still only support the use of 64-bit registers, we
recognize RAX, but not EAX, AX, AH, or AL.)